Always rerun the integTest task from IDEA
When starting a test from IDEA, you really want
to run it and expect feedback on the tests which
executed. This change now detects if the Build
has been started from IDEA and then reruns the
test.

It also adds a `--no-rerun` command line option,
so the rerunning can be switched off.