Fix verification file generation
It was possible that the verification file generation failed if
one or more artifacts where missing from the local dependency cache
(most likely due to manual deletion of artifacts).

This commit fixes the problem, making sure that we capture the
exception. The generated file will miss the entry corresponding
to the missing file, but a warning is visible in the console.

Fixes #12260