Ensure that an exception thrown while initializing an worker process to execute worker API actions does not go missing.
Previously, these were caught by `WorkerAction` so that it could attempt to send the failure as the result message for the first request. However, this kind of failure also means that the serializers used for the connection are not initialized correctly and so no messages could be received by the worker process. This means that the failure would never be sent to the build process.

This change simply propagates the failure so that it ends up being logged to the worker process' stderr and is then forwarded to the Gradle logging system. There's no attempt to notify the build process of the specific failure (the process' exit code is not 0, so the build process knows something has gone wrong) or to gracefully clean up the connection. The assumption here is that this kind of failure is quite rare and is a result of some programming error in Gradle.