Reimplement source to class name mapping for Java
Previously, mapping from source files to class names was done by extracting
the class names from the class files names, which wasn't 100% accurate.
The Groovy incremental compiler, on the other hand, was implemented using
an AST transformation which could extract the class names from the compiler
itself, therefore giving an accurate mapping from class names to source files.

The same is now done for the Java compiler, using a Java compiler plugin.
This fixes several issues where we could not figure out properly what files
to recompile when the name of the class being used in a dependency wasn't
directly related to the source file name.

However, this introduces a problem when the compiler is not a proper JDK
compiler. In particular, if compiling using a "path to compiler", the
plugin cannot be loaded and incremental compilation will fail.