'mainClass' and 'mainModule' properties for application plugin extension
This adds the new property 'mainModule' to define a module that should
be executed if the applications runs with a --module-path.

It adds the 'mainClass' property as replacement for the 'mainClassName'
getter/setter. The behavior is also extended: The main class is now
automatically added to the module-info.class (through compileJava task)
and the Jar's manifest (through jar task).