Avoid eager creation of module identifier strings
and optimize `DefaultModuleIdentifier#toString`. This
showed up in a profiler snapshot in IDE sync: apparently this
code is invoked a lot. Because the calls to the loggers are not
guarded, the string representation is always created even if
logging is disabled.

The `toString()` method was using `String.format` which is
particularly slow.