Avoid unnecessary intersection computations
In the `gradle/gradle` build, all calls to _intersect_ were done
on the same sets of modules. Because we create immutable copies
every time, those were different instances.

By just checking, when we compute the _union_, that the sets are
the same, we avoid 100% of intersection calls in the `gradle/gradle`
case!