WrapperHttpsIntegrationTest: Test (authenticated) download via a proxy
The "downloads wrapper via authenticated proxy" test is currently
expected to fail to demonstrate the issue and will be fixed in the next
commit.

Note that as of JDK 8u111 the basic authentication scheme is disabled by
default when proxying HTTPS [1]. Do not disable any schemes by setting
the according system property to an empty string.

[1] https://bugs.openjdk.java.net/browse/JDK-8168839?focusedCommentId=14019675&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-14019675

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>