Make start script generation depend on the actual classpath
We need to analyze the content of the classpath file collection to
decide which jars go on the module path and which go on the classpath.
Therefore the content is now an input again (and not only the names
of the jars).

This reverts the change in #1936 that removed the classpath content
as an input.