Change PluginsBlockSourceWrapper to only override the relevant changed method
Instead of decorating a ScriptSource, override TextResourceScriptSource and override the only method that needs changing. This is in order to make it more evident why the override is there. Reason being the class name returned - the script is compiled in two phases - plugins block, then body. The generated artifacts have the same class name and they clash when packaged in a single Jar.