Fix handling of configuration level excludes
For realised Ivy metadata, excludes at the level of configurations added
by rules need to be explicitly serialized as we cannot rely on the
usual derivation from the base Ivy metadata since the mapping between
the added configuration and the base one is lost.

The test has also been adapted to show that one derived variant excludes
while another does not.