Modular app: 'run' task uses application jar instead of classes
By default, Java expects everything that belongs to one module to be
in the same location (Jar or folder). Since in Gradle resources
or outputs from different compilation steps are in separate folders,
everything that is not in the same 'classes' folder as the
module-info.class becomes unaccessible. Hence we always use the complete
Jar now to run a module via the :run task.