Run JVM test classes in a reproducible order
This addresses part of #8520.

Ideally, we would support running test classes in a stable but
non-obvious ordering that doesn't encourage a reliance on the
order in which tests are run. The ordering here (which is easier
to achieve) is stable but also obvious. This is still preferable
to having non-deterministic ordering; it fixes the issue of having
a set of tests pass pre-merge CI for one commit and then fail in
other commits due to running in a different order.

Signed-off-by: Alex Landau <alandau@palantir.com>