Apply the Java derivation strategy to the Java Platform plugin
This commit makes sure that the Java Platform plugin also applies the
Java derivation strategy, as it, as it names indicates, also belongs
to the Java ecosystem. Not doing so triggers some weird ordering issues
if another project, which is a Java library, tries to resolve the same
dependencies as the Java platform, but after the platform has resolved
them.

Because of component metadata caching, the first project to resolve
a dependency will be the "source of truth" for dependency metadata for
the module for the whole build. It means that if it was a Java library,
then the derivation strategy was applied and the project could resolve
everything properly. However, if the first project to trigger resolution
of a module was a Java Platform, then we would cache component metadata
_without_ derivation. This means, in particular, that platforms wouldn't
be derived for external Maven modules.

This explains why in #12728, order of execution of tasks matters: if
the subproject was resolved _first_ (gradle sub:compileJava resolve)
then everything was fine, but if the platform was resolved first, then
the rest of the build would fail.

The fix is to apply the derivation strategy to the Java platform plugin
too. However, in an ideal world, we should refactor the Java plugins so
that they consist of several layers:

- a "schema" layer, configuring the ecosystem schema, the derivation strategy, ...
- a "base" layer, consisting of what the Java Base plugin does today
- a "platform", directly on top of "schema"
- an application, and a "library", on top of "base"

Fixes #12728