Tell JaCoCo to ignore kotlin-dsl generated files
As stated in the issue #10956 the code generated by Gradle affects severely to our coverage.

I've created the annotation at org.gralde.api.internal because I expect no one uses it, it can be named differently while it contains Generated in the name
I haven't added a test checking this explicitly, no idea where to do it properly, but checked on out projects and there is still a small issue, accessors that contain methods with default arguments are still shown as not covered.

Signed-off-by: CristianGM <cristiangm@gmail.com>
Signed-off-by: cristian-garcia <cristian.garcia@scmspain.com>