Consider the derivation strategy when requesting component metadata
Gradle has a single, build scoped cache for dependency metadata. It's
using a build cache scope because component metadata is quite expensive
to read and process. However, component metadata _may_ be different
based on the plugins applied on a project. For example, Java projects
apply what we call a "Java derivation strategy", which is capable
of transforming regular POM files into so-called "platforms" but
more importantly, they are what map the different Maven scopes to
actual "variants" in variant-aware dependency management.

The problem is that it's not a requirement that all projects are
Java projects. It is possible to have native subprojects, or projects
which do _not_ apply the Java base plugin. While it's in general an
error to have a project which wants to perform "Java dependency
resolution" not to apply the Java derivation strategy, there are
consequences in not doing so. In particular, Gradle becomes order
dependent: if a subproject which does not apply the derivation
strategy resolves a module first, then subsequent resolutions for
projects which _do_ apply the derivation strategy would get wrong
metadata.

To avoid this, this commit makes sure that when we consume module
metadata, we actually use the derivation strategy as part of the
processed metadata cache key.

This does _not_ fix the fact that the consumers _should have_ applied
the derivation strategy and will make realizing this harder, but it
_does_ fix the inconsistency in resolution, which is probably a
much bigger issue (for caching and reproducibility).