Change the strategy for handling undeclared system property reads when instant execution is enabled to allow reads of a property whose value is `null`.
Also treat system properties read in this way as build inputs so that when the value changes (to non-null) then configuration happens again and the validation is applied.

This is intended to help plugin authors migrate to instant execution, when the plugin uses some library to do some work at configuration time.