Add support for include files for antlr-plugin
Adds a new `@Input` property `Set<Pattern> includeFilesPatterns` to `AntlrTask` to contain the regex patterns to match the include-files (applied on the project-relative-path of each input file). If `includeFilesPatterns` is not empty, "incremental grammar builds" are disabled, as there is no way to know which include file is used in which grammar. Include-files are not passed to the antlr tool.

fixes #13005

Signed-off-by: Robert Stupp <snazy@snazy.de>