Fix metadata realization
The "dependency filter" in module metadata is a performance optimization,
which allows filtering out some particular kinds of dependencies _without
having to duplicate the collection_ which is particularly expensive.

However, when we call the builder to mutate metadata, we shouldn't use
the current value of the builder as the initial value, because it causes
the mutated state to be in a wrong state.

This fixes a failing test in "force realize" case.