Don't write in RO dependency cache when parent POM is missing
This commit works around a bug whenever the read-only dependency cache
used to contain a parent POM, but it has been cleaned up in between
(for example via the seeding build).

In this case, if a "consumer" build uses the read-only cache and
resolves a dependency which has the same parent POM, it will attempt
to write in the read-only dependency cache even if it shouldn't.

The ideal fix would be to avoid tweaking the failsafe cache, but it
appears that some read operations _may_ trigger a write operation
in between (as a cleanup side effect). This commit therefore makes
sure that we don't even attempt to call a write operation in this
case.

Fixes #12996