Proposal: Add opt out for "no valid plugin descriptors" noise
In multiple projects, I've found myself using buildSrc for shared build logic rather than solely just gradle plugins. If this is done though, Gradle logs noise about finding no plugin descriptors. While this would make sense if this was programmer error, it's often not.

This PR is a proposal to add a property to disable this log.