Update behaviour for exclude upper bound in range
The upper bound of a version range, when it is an exclusion, now acts as
a smart prefix. Versions that were excluded before remain excluded. In
addition, versions that start with the upper bound are also excluded.
This resolves the unexpected case where 2.0-dev1 is included when the
upper bound is 2.0[.

Fixes #13050