Add dedicated placeholder assertion error type
This commit introduces a dedicated placeholder assertion error
type which is used whenever a subtype of `AssertionError` cannot
be deserialized. Before, we would use the regular `PlaceholderException`,
which happens to discard the fact that it's an `AssertionError`.
The consequence is that IntelliJ IDEA in combination with Spock can't
infer the correct failure kind, and displays a wrong icon (see #12302).

It does not, however, fix the problem that the "Click here to see diff"
link doesn't show up. The reason is that this feature builds on top
of a dedicated subtype of `AssertionError` that the messaging bus
isn't aware of (org.opentest4j.AssertionFailedError and its legacy JUnit
equivalent).

Fixes #12302