Introduce an API for variant dependency substitution
This commit introduces a new API to deal with _variant-aware_ dependency
substitution. Before this change, it wasn't possible to substitute a
dependency on a platform, for example. This change allows configuring
the attributes of a dependency, both on the "source" side (what dependency
we need to substitute) and the "target" side (what dependency is going to
be used instead).

Some DSL short-hand notations are added for common use cases.

Fixes #13204