Simplify unnecessary complex Soak test setup
We already have a separate project with special handling on CI.
We don't need an additional Category and a special DistributionTest
type.

We can now just use the default 'integ test' setup inside the :soak
project and do not need to create additional configurations for
soak tests anymore in the build script.