Do not add commit and branch name as input to ALL tests
This breaks the cache every time the commit hash changes.
We now only add the branch for performance tests and read the
commit from the current Gradle version. It is stored in the
build receipt packaged in base-services.jar (which is ignored
as input already by a classpath normalization rule).