Move release-feature.txt to :launcher
So we do not need to ship an almost empty docs.jar anymore.
It is for the launcher and the testing was already done in the
launcher project. There is also no need to copy the file around
from specific locations. We place it in the default 'main/resources'
folder now.