Include all projects that are the target of a resolved project dependency in the configuration cache.
This allows the correct workspace to be calculated for an artifact transform of a file dependency, regardless of whether the file dependency is the output of a task or not. This in turn means that the transform will be up-to-date on the first read from the cache (assuming nothing else has changed).