Do not eagerly execute artifact transforms for any file dependencies when writing to the configuration cache.
These files may change and so should be treated as configuration cache inputs if the result is eagerly cached. By lazily executing these transforms, these files do not need to be tracked as inputs to the configuration cache and the up-to-date checks for these transforms can happen in parallel during execution rather than at the start of the build.

Transforms for external remote dependencies are still run eagerly when writing the cache entry.