Do not publish normalized version of Tooling API Jar
This is declared as a separate publication to remove the timestamp
from the version. Gradle automatically creates a publish task
for this version. This should not be executed with the 'publish'
lifecycle task and hence we deactivate it.

We also no longer provide the local version for non-shaded Jars
where it is not used. Having the publication with a different version
leads to the following error during metadata generation:
'not able to resolve a dependency on a project with multiple
publications that have different coordinates.'