TestKit must use a user home for the reduced distribution it runs with
If a test that uses TestKit does not have the 'full' distribution
available, it needs to use the user home dedicated to the distribution
it runs on. And not the default shared test-kit home in the system
temp folder. Otherwise, multiple TestKit tests based on different
distributions might end up sharing a daemon.