Fix Java attribute describer when category is missing
It's possible, during variant matching, that we match on the
category attribute with value "documentation", but that we
fail on the "docsType" attribute. In this case, when displaying
what didn't match, we want to make sure we can display a
reasonable error message. Currently because the "category" matched,
we just ignore the "docsType".