Add some hand-crafted serialization for `PublishArtifactLocalArtifactMetadata` types to the configuration cache.
This is required because instances of this type use themselves as their own id, and so when serialized as an id (which is all that is required) drag in a lot of unnecessary and unserializable state.

A better change would be to separate out an immutable id from `PublishArtifactLocalArtifactMetadataCodec`, however the Eclipse tooling model builder currently takes advantage of the coupling to navigate from the id to the producing task for the artifact. Fixing this is too large a change for now and something that can happen later.