Support more notations for `MavenPublication#artifact()`
Currently, there's no good way for a user to have a lazy artifact
provider, even if the producer is lazy (e.g, a `TaskProvider` or
a `Provider<RegularFile>`.

This commit fixes that by supporting more notations in the converter
and automatically using `LazyPublishArtifact` when a provider is
supplied.

The lazy publish artifact now also supports more, in particular it
can accept any task as long as it provides a _single_ output file.

Fixes #7958