Introduce `@InternalApi`
If a package is annotated with `@InternalApi`, this package and
all the subpackages it contains are not exposed in the Gradle API
jar.

Similarly, if a class within a non-annotated package is itself
annotated with `@InternalApi`, then it's not visible in the
Gradle API jar.

This approach lets us introduce new internal APIs which are
truly invisible to users, without relying on Java modules.

For now, because we only care about the Gradle API jar, it
means that the internal APIs are not visible in `buildSrc`,
but they are effectively usable in build scripts.

See #13501