Write the artifact transforms for external dependencies to the configuration cache as a spec, rather than eagerly running the transform and writing the resulting files to the cache.
This means that no transforms are run eagerly. All transforms may take task outputs or other changing files as parameters and it is not easy to determine whether or not this is the case. So, for now assume that all transforms have changing parameters. A later improvement could be to eagerly run transforms for remote artifacts that do not have changing parameters.