Isolate and simplify code for determining the current version and name
This is now located in 'module-identity', defining a plugin that is
individually applied by all subprojects.

Groovy code has been translated to Kotlin, duplication has been removed.

The build receipt creation is now handled by the baseServices project,
as that is the place, where it is packaged in the end.