Introduce a `JVM` extension
This commit refactors how the JVM plugin services are wired, so
that they are configured at a higher level. It basically introduces
a new plugin, the `jvm-ecosystem` plugin which is the top-level
base plugin for all JVM plugins.

This plugin configures the `jvm` extension, which gives access
to the JVM services as well as creates the minimal setup for JVM
projects, that is to say configure the schema, derivation rules
and create the source set container.