Remove setting ignoreIncomingBuildReceipt based on requested tasks
This introduces timing to build configuration, as it needs to be
set early. It is better not to rely on this, but set such properties
explicitly when invoking Gradle from the command line via '-P'. In this
case, this is already done in `CompileAll.kt` as part of the TeamCity
configuration:

https://github.com/gradle/gradle/blob/ba7596186caea80ac56883c33c634332cc7c10b4/.teamcity/Gradle_Check/configurations/CompileAll.kt#L24