Do not store metadata rules in the in-memory cache
This commit avoids storing the component metadata rules when
we store component metadata in the in-memory cache. While this
shouldn't be necessary in practice because the in-memory cache
only lives for a build, it's possible that the instance is
actually only freed once we reach enough memory pressure, which
means we can unnecessarily keep project references in memory.

Fixes #13497