Handle corner cases for excludes
When simplifying an AnyOf we could end with a two elements AnyOf where
one of the elements is also an AnyOf. In that case we missed the
flattening and with the right setup it can cause exclude computation to
blow up performance.
This change recognizes this corner case and special cases for it.