Move watchFs logging to build operations
so we don't print the excessive logging at the
start and the end of the build.

While there are now build operations which have
the watching statistics, is is hard to capture
those build operations:
- The watching started operation is so early, no
  build logic did run at that point so we can't
  capture it.
- The watching finished operation is so late, that
  build services already did shut down, so a
  BuildOperationListener - which doesn't work with
  configuration caching - is the only way to capture
  them.

Since we can't capture the start operation, the
soak test can't assert anything about the received
events during the build any more.