Fix missing category attribute on plugin classpath
The plugin compile classpath configuration was missing the `category`
attribute. The consequence is that if a plugin was published with
Gradle metadata, and that for some reason one of the attributes like
the target JVM version didn't match on the library variants, then we
would fallback on the "sources" variant because it has less attributes
and is compatible with all the requested attributes!

Fixes #13659