Allow toolchains to be downloaded in parallel
Reworked to use a striped lock for each toolchain individually. This
allows parallel downloads of jdks. Removed the need for the read only
lock by using marker files that are marking provisioned jdks.
If the same toolchains happens to provisioned in parallel, once the
second builds acquires the lock, it will reuse the provisioned toolchain
instead of downloading and unpacking again.