Appply only required plugins to architecture-test subproject
The idea is to dissect what this subproject really is and then possibly extract that to a separate convention plugin. Currently it states that it is a java library which it is not. The complication made by this statement is that in lifecycle plugins we have to filter out this project and conditionally apply additional logic to the projects that are in fact java libraries.