Introduce experimental feature to exercise Gradle 7 removals
This commit introduces a new experimental "feature" which can
be used to preview what's going to happen in Gradle 7 with some
removals. In this case, this commit removes a number of configurations
like `compile` and `runtime` which have been deprecated for
quite some time already, and makes the `maven` plugin a no-op.