Reimplement `available-at`
This commit reworks how `available-at` is implemented in Gradle
Module Metadata. Previously, if a module was referencing a variant
using `available-at`, then the variant of the module would add
the "referenced" module as a dependency. That module _could_ have
a variant with the same attributes but it wasn't enforced.

Because it was added as a _dependency_, it also means that the
behavior of resolution depended on whether resolution was transitive
or not. It also means that the _actual_ variant couldn't be mutated
independenly of the referenced module.

This caused a number of issues, in particular the one referenced
in #14017.

The new implementation works by actually parsing the referenced
module metadata and _inlining_ variants. For backwards compatibility,
we fallback to the old behavior if we can't find the target module
or a matching variant.