Simplify snapshotting service
After discussing this with Lorant, we decided to base the implementation of
the snapshotting service to use the FileSystemAccess class. While snapshots
can be retrieved from FileSystemAccess, when considering file normalization,
we should instead return finger prints. But since it is not completely clear
yet how test distribution should use this, we are not introducing this for
the time being.