Fail if the referenced module doesn't exist
For reproducibility, it's better to _fail_ if a referenced module
isn't published whenever we resolve a module which has an available
at component: otherwise it means that depending on _when_ we resolve,
the cached metadata would be different.

Similarly, we make sure that the inlined variants are immutable
by testing if the module is changing.