Fix behavior of `available-at`
This commit removes the check for `changing`, which will not be
an error as it's preventing a usage scenario where a multi-project
published with `available-at` relationships _and_ using a snapshot
version would be broken.

Instead we _trust_ users that they wouldn't publish a release module
which depends on snapshots.

In addition, some tests have been removed because they exercise
a scenario which isn't supported anymore: it's not possible to
consume a module which has been "partially published".