Pull authoring multi project builds topic to its own section in docs nav menu
So that it is visible directly below "Learning the basics" menu.
Currently it this important topic is buried within "Learning the basics" close to the bottom.
Next steps after pulling it up could be to create a new hamburger menu for the multi-project build authoring and split the lengthy page into multiple sub-menus.