Fix project dependency substitution in generated metadata files
This commit adds handling of project dependency substitutions to
other projects. Previously, if a project was substituted with
another project, then when publishing resolved versions, the
substituted project wouldn't be replaced with the target project
in the generated metadata files (pom.xml, ivy.xml and Gradle
Module Metadata). While this is certainly an edge case that
shouldn't be encouraged, this properly handles the bug.

Fixes #14039