Introduce a lazy GUtil version of elvis
The GUtil.elvis method had the drawback of requiring
to provide a _value_ in case the tested value wasn't null.
In short, it works differently from an `if` because it forces
the compute the "else" value even if we never need it.

This commit introduces `getOrDefault` which uses a lambda
for the "default" value instead, which is lazy.

It also fixes the contract of `elvis` which could effectively
take a nullable as the default value _and_ return `null`.

Fixes #14237