Fix handling of deleted files when using verification
Previously, dependency verification was failing whenever an artifact
wasn't found in the local cache. This caused problems for artifacts
and metadata deleted legitimately by the automated cache cleanup
process.

This commit makes sure that if an artifact is deleted, we, instead,
try to redownload it and verify the result of downloading, as if
it was a newly downloaded artifact.

Fixes #12713