Add infrastructure to run split performance tests
This only creates the buckets, the code to make this
actually work in the Gradle build itself is still
missing.

We need to add the buckets to `master`, so we can
try out the build logic on a branch, since Teamcity
does not support adding build types or dependencies
on branches.