Introduce API to declare consistency between configurations
This commit introduces a new method on the `Configuration` interface,
which lets the user declare that it should resolve consistenly with
another configuration.

The relationship between two configurations is directed. In other words,
"A must resolve consistenly with B" means that the reference is B.
When A is going to be resolved, we start with resolving B, then inject
strict dependencies on the resolved versions of B into A.

For example, `runtimeClasspath.resolveConsistentlyWith(compileClasspath)`
means that the compile classpath is the reference. That is to say, you
expect that the versions of dependencies you use at compile time will
be the ones found at runtime.

There's only one "source" of consistency for a configuration. It is
expected that the consistency source is also resolvable, but it doesn't
have to use the same attributes for resolution. In the example above,
we resolve the "compile classpath" which would typically resolve using
the usage=JAVA_API attribute, collect the resolved dependency versions,
then inject the result to the runtime classpath configuration as
constraints, but runtime classpath would use usage=JAVA_RUNTIME.

Because we use strict dependencies under the hood, it means that:

- a conflict on first level dependencies would trigger a build failure
- a transitive dependency can be downgraded if the reference version
is lower

In any case, we're using regular constraints, which means that all
selectors are used during resolution, and that an incompatible selector
would trigger an error that the user would have to fix.