Increase heap size for ScriptCachingIntegrationTest
We see a lot of flaky `in-memory script class loading cache releases memory of unused entries`
which expects the class to be loaded from in-memory cache but actually cache-missed.

After analyzing the daemon log, it seems like that sometimes daemon is shut down without finishing
all iterations because of:

...
[DEBUG] heap: GC rate: 2.67/s heap usage: 98%
[WARN] Expiring Daemon because JVM heap space is exhausted
LIFECYCLE] Daemon will be stopped at the end of the build after running out of JVM memory
...

Thus, there're actually different daemons in the build, resulting the flakiness.

In the test, there're 256-64=192M heap for Gradle daemon, which seems not enough.
Now we increase the test daemon size to 512M.