Avoid construction of full component metadata for included projects (#14627)
This causes the full metadata, including all variants, to be built
eagerly which may cause task realization when the attributes for
the variants are calculated. For example: artifact type is retrieved
from AbstractArchiveTask, Java target version is retrieved from
JavaCompile task.

The only reason to construct the metadata here was to obtain the
module ID. We can shortcut by calculating it directly based on
the configured project's state.