Fix snapshotting file trees with filtered out roots
A file tree can have a regular file or a missing file
as a root. The include/excludes apply to those roots,
so it is possible that we filter out the root of the
file tree.

DirectorySnapshotter.snapshot() is able to walk missing
files and regular files, more or less by accident.
Though when those files are filtered
it returns `null`, causing an NPE when we try to store
the result in the VFS.

Instead of going through DirectorySnapshotter for things
which are not directories, we stat the file first and use
DirectorySnapshotter only for directories.