Spike central declaration of dependencies with TOML
This commit introduces a TOML file which is used to declare a number
of "raw" dependency coordinates. Each declaration is given an alias,
which is then used to generate accessors which are available in every
project.

For example, given the following declaration in the TOML file:

```
[dependencies]
groovy = { group = "org.codehaus.groovy", name = "groovy", require = "3.0.2" }
```

then in a build script one can add a dependency on groovy just by
doing:

```
dependencies {
   api(libs.groovy)
}
```

This works by generating an extension at configuration time. The advantage
is that now all projects can "align" on what declaration to use.
It doesn't make sure that the _resolved_ versions would be the same in
every project, but it makes sure that they declare the same thing.

The file also supports the notion of "bundle", which is basically a number
of dependencies which are always grouped together.