Introduce a lazy field factory
This commit introduces a `Lazy` type, which aims at making it
easy to create final fields which initialization is costly and
would benefit from being deferred.

The `Lazy` type has 3 factory methods to make it easier to
choose between 3 types of synchronization depending on the
use cases. The simplest implementation isn't thread-safe.

This code was originally written by me for a 3rd party library:
https://github.com/melix/rigel