Introduce repositories in settings
This commit adds the ability to declare convention repositories in
the `settings.gradle(.kts)` file. Repositories which are declared
in settings act, by default, as conventions. That is to say that if
a project doesn't add repositories of its own, then the repositories
declared in settings will be used.

Effectively, this is a replacement for the common pattern:

```
allprojects {
   repositories {
      ...
   }
}
```

It's worth noting that `buildSrc`, in particular, won't "see" those
repositories as it is considered as an included builds. Included
builds will use their own repositories (and if themselves declare
settings repositories, those are used when building the included
project).

If repositories need to be shared between the main build and the
buildSrc build, then a settings plugin needs to be published. It's
a current limitation of Gradle that settings plugin cannot be
included as regular composite builds, because of evaluation ordering
issues.