Make it possible to prefer settings repositories for resolution
This commit adds the ability to select what repositories are going
to be used during project dependency resolution. By default, whatever
repositories the project declares are used, shadowing the repositories
declared in settings.

However, it is possible to configure Gradle so that it always uses the
settings repositories. In this case, the build author has two options:

- warn if a project (or a plugin in a project) declares a repository
- fail if a project (or a plugin in a project) declares a repository

This makes it possible, for a build author, to make sure that if they
want global consistency, the repositories used by any project are
always the same.