Ensure buildSrc/build.gradle.kts is not cached between tests in compile avoidance test
Having it cached causes tests that use the daemon reuse it from in-memory cache (because runtime classpath is always the same for buildSrc) which as a result does not store it in the test-local script cache directory. When the test connects to a different daemon (or starts a new daemon), it gets an in-memory cache miss and recompiles the script. If this happens during the second Gradle invocation when we test for compile avoidance, this situation breaks the assertion, making the test flaky.