Let Kotlin DSL script compilation disable compatibility mode for kotlin new inference engine
DisableCompatibilityModeForNewInference is a special flag to preserve compatibility with the old type inference.
It only appeared in 1.4 and is turned off by default.
What it does is artificially lower the priority to resolve in some corner cases
(for example, some cases with SAM conversions, as in our case).
In the old inference, such calls were not prioritized properly.
Now this flag was introduced to smooth migration.
A funny fact is that when the -XXLanguage: +NewInference flag is passed from the command line, Gradle or IDE,
the DisableCompatibilityModeForNewInference flag is turned on automatically,
but not in the case of embedded Kotlin in Gradle,
since the language version settings are manually configured there, bypassing this logic.

Signed-off-by: Paul Merlin <paul@gradle.com>