Put project type safe accessors generation under experimental flag
Type-safe project accessors are not the focus of the 6.8 release.
We still want feedback, but there are some open questions, like
the relationship with composite builds, which prevent us from
enabling this by default.

As a consequence, early adopters can set the TYPESAFE_PROJECT_ACCESSORS
feature preview to enable the generation of such accessors, with
some restrictions on the pattern of project names.

Use at your own risk.