Detect task cycles caused by finalizedBy nodes
When we walk the task graph to detect cycles, we follow the dependencies of a given node. If A dependsOn B, then the graph is walked from A to B.

Finalizer nodes have an inverse relationship with the nodes they finalize when compared to nodes that have a dependency relationship.
If A is finalizedBy B, then A is a finalizingSuccessor of B. When the graph is walked from A following the successor nodes, B is not visited, unless it happens to appear in the graph when following the dependency order.

If while computing the execution plan we found a cycle, we know that the offending node is reachable from itself following its dependency (successor) edges.
Walk the graph from this offending node instead of starting from the original start node (which can miss finalizer nodes due to inverse dependency relationship).