Simplify Git instructions for signing patch sets
`git filter-branch` has a plethora of pitfalls. Regardless, the existing
recommended command in the CONTRIBUTING file was misleading at best,
plain wrong at worst: If written as-is, the full history of the project
would be rewritten and appended with the s-o-b trailer.

Replace it with a simpler invocation of `git rebase --signoff` which is
supported since at least Git 2.13 (May 2017). Using `git rebase` also
has the added benefit of not adding the line if it already exists for a
commit.

Signed-off-by: Daniel Knittl-Frank <knittl89+git@googlemail.com>