Allow the creation of multiple dependency models
This commit refactors the code so that a user can create multiple dependency
models. The rationale is that we may want, for example, to separate the main
dependencies from the test dependencies. We may also want to import the model
from different _platforms_ which would potentially clash and provide the
same aliases.

By namespacing, we can have build scripts reference different models:

```
dependencies {
    implementation(libs.foo)
    implementation(spring.framework)
}
```